Revert "v1.0.7 (#41)"
This reverts commit b26342b14177c3aaf8afc019ceec08c715780c25.
