+++ cmd/libgen-cli/search.go
@@ -67,10 +67,6 @@ var searchCmd = &cobra.Command{
 		if err != nil {
 			fmt.Printf("error getting output flag: %v\n", err)
 		}
-		publisher, err := cmd.Flags().GetString("publisher")
-		if err != nil {
-			fmt.Printf("error getting publisher flag: %v\n", err)
-		}
 
 		// Join args for complete search query in case
 		// it contains spaces
@@ -86,7 +82,6 @@ var searchCmd = &cobra.Command{
 			RequireAuthor: requireAuthor,
 			Extension:     extension,
 			Year:          year,
-			Publisher:     publisher,
 		})
 		if err != nil {
 			fmt.Printf("error completing search query: %v\n", err)
@@ -141,19 +136,19 @@ var searchCmd = &cobra.Command{
 			Keys: &promptui.SelectKeys{
 				Next: promptui.Key{
 					Code:    readline.CharNext,
-					Display: "↓ (j)",
+					Display: "↓",
 				},
 				Prev: promptui.Key{
 					Code:    readline.CharPrev,
-					Display: "↑ (k)",
+					Display: "↑",
 				},
 				PageUp: promptui.Key{
 					Code:    readline.CharForward,
-					Display: "→ (l)",
+					Display: "→",
 				},
 				PageDown: promptui.Key{
 					Code:    readline.CharBackward,
-					Display: "← (h)",
+					Display: "←",
 				},
 			},
 		}
@@ -175,9 +170,9 @@ var searchCmd = &cobra.Command{
 		}
 
 		if selectedBook.Author == "" {
-			fmt.Printf("Download starting for: %s by N/A\n", selectedBook.Title)
+			fmt.Printf("Download started for: %s by N/A\n", selectedBook.Title)
 		} else {
-			fmt.Printf("Download starting for: %s by %s\n", selectedBook.Title, selectedBook.Author)
+			fmt.Printf("Download started for: %s by %s\n", selectedBook.Title, selectedBook.Author)
 		}
 
 		if err := libgen.GetDownloadURL(&selectedBook); err != nil {
@@ -213,6 +208,4 @@ func init() {
 		"libgen-cli to save your download.")
 	searchCmd.Flags().IntP("year", "y", 0, "filters search query results by the "+
 		"year provided.")
-	searchCmd.Flags().StringP("publisher", "p", "", "filters search query "+
-		"results by the publisher provided")
 }
