+++ go.mod
@@ -5,7 +5,7 @@ go 1.14
 require (
 	github.com/cheggaaa/pb/v3 v3.0.4
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
-	github.com/dustin/go-humanize v1.0.1-0.20200219035652-afde56e7acac
+	github.com/dustin/go-humanize v1.0.0
 	github.com/fatih/color v1.9.0
 	github.com/manifoldco/promptui v0.7.0
 	github.com/spf13/cobra v0.0.7
