+++ libgen/api.go
@@ -33,49 +33,6 @@ import (
 	"github.com/fatih/color"
 )
 
-// Book is the struct of resources on Library Genesis.
-type Book struct {
-	ID          string
-	Title       string
-	Author      string
-	Filesize    string
-	Extension   string
-	Md5         string
-	Year        string
-	Language    string
-	Pages       string
-	Publisher   string
-	Edition     string
-	CoverURL    string
-	DownloadURL string
-	PageURL     string
-}
-
-// SearchOptions are the optional parameters available for the Search
-// function.
-type SearchOptions struct {
-	Query         string
-	SearchMirror  url.URL
-	Results       int
-	Print         bool
-	RequireAuthor bool
-	Extension     string
-	Year          int
-	Publisher     string
-}
-
-// GetDetailsOptions are the optional parameters available for the GetDetails
-// function.
-type GetDetailsOptions struct {
-	Hashes        []string
-	SearchMirror  url.URL
-	Print         bool
-	RequireAuthor bool
-	Extension     string
-	Year          int
-	Publisher     string
-}
-
 // Search sends a query to the search.php page hosted by gen.lib.rus.ec(or any
 // similar mirror) and then provides the web page's contents provided from the
 // resulting http request to the parseHashes() function to extract the specific
@@ -105,8 +62,29 @@ func Search(options *SearchOptions) ([]*Book, error) {
 	q.Set("column", "def")
 	options.SearchMirror.RawQuery = q.Encode()
 
-	b, err := getBody(options.SearchMirror.String())
+	// Execute GET request on search query
+	client := http.Client{
+		Timeout: HTTPClientTimeout,
+		Transport: &http.Transport{
+			Proxy:           http.ProxyFromEnvironment,
+			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		}}
+	r, err := client.Get(options.SearchMirror.String())
 	if err != nil {
+		return nil, err
+	}
+	if r.StatusCode != http.StatusOK {
+		return nil, err
+	}
+
+	// Read body of response to get HTML
+	b, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	// Close request and handle possible error.
+	if err := r.Body.Close(); err != nil {
 		return nil, err
 	}
 
@@ -120,7 +98,6 @@ func Search(options *SearchOptions) ([]*Book, error) {
 		RequireAuthor: options.RequireAuthor,
 		Extension:     options.Extension,
 		Year:          options.Year,
-		Publisher:     options.Publisher,
 	})
 	if err != nil {
 		return nil, err
@@ -143,8 +120,21 @@ func GetDetails(options *GetDetailsOptions) ([]*Book, error) {
 		q.Set("fields", JSONQuery)
 		options.SearchMirror.RawQuery = q.Encode()
 
-		b, err := getBody(options.SearchMirror.String())
+		client := http.Client{
+			Timeout: HTTPClientTimeout,
+			Transport: &http.Transport{
+				Proxy:           http.ProxyFromEnvironment,
+				TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+			}}
+		r, err := client.Get(options.SearchMirror.String())
 		if err != nil {
+			log.Printf("error reaching API: %v", err)
+			return nil, err
+		}
+
+		b, err := ioutil.ReadAll(r.Body)
+		if err != nil {
+			log.Printf("error reading response from API: %v", err)
 			return nil, err
 		}
 
@@ -169,19 +159,65 @@ func GetDetails(options *GetDetailsOptions) ([]*Book, error) {
 				continue
 			}
 		}
-		if options.Publisher != "" {
-			if !strings.Contains(book.Publisher, options.Publisher) {
-				continue
-			}
-		}
 		if options.Print {
-			if err := printDetails(book); err != nil {
+			var fsize string
+			size, err := strconv.Atoi(book.Filesize)
+			if err != nil {
+				fsize = "N/A"
+			} else {
+				fsize = humanize.Bytes(uint64(size))
+			}
+
+			// Print separation lines
+			fmt.Println(strings.Repeat("-", 80))
+
+			// Print ID + Title
+			fTitle := fmt.Sprintf("%5s %s", color.New(color.FgHiBlue).Sprintf(book.ID), book.Title)
+			fTitle = formatTitle(fTitle, TitleMaxLength)
+			if runtime.GOOS == "windows" {
+				_, err = fmt.Fprintf(color.Output, "%s\n    ++ ", fTitle)
+				if err != nil {
+					return nil, err
+				}
+			} else {
+				fmt.Printf("%s\n    ++ ", fTitle)
+			}
+
+			// Slice author name if it exceeds AuthorMaxLength
+			var formatAuthor string
+			if len(book.Author) > AuthorMaxLength {
+				formatAuthor = book.Author[:AuthorMaxLength]
+			} else if book.Author == "" {
+				formatAuthor = "N/A"
+			} else {
+				formatAuthor = book.Author
+			}
+
+			err = prettify("author", formatAuthor, color.FgYellow, "-25")
+			if err != nil {
 				return nil, err
 			}
+			err = prettify("year", book.Year, color.FgCyan, "4")
+			if err != nil {
+				return nil, err
+			}
+			err = prettify("size", fsize, color.FgGreen, "6")
+			if err != nil {
+				return nil, err
+			}
+			err = prettify("type", book.Extension, color.FgRed, "4")
+			if err != nil {
+				return nil, err
+			}
+			fmt.Println()
 		}
 
 		// Add valid book to the []Book for the search
 		books = append(books, book)
+
+		if err := r.Body.Close(); err != nil {
+			return nil, err
+		}
 	}
 
 	return books, nil
@@ -213,9 +249,12 @@ func GetWorkingMirror(urls []url.URL) url.URL {
 
 	for {
 		randMirror := urls[rand.Intn(len(urls))]
+
 		if CheckMirror(randMirror) == http.StatusOK {
 			mirror = randMirror
 			break
+		} else {
+			continue
 		}
 	}
 
@@ -225,7 +264,7 @@ func GetWorkingMirror(urls []url.URL) url.URL {
 // ParseDbdumps takes in a HTTP response and scans it for
 // any string that matches a filepath and returns all results.
 func ParseDbdumps(response []byte) []string {
-	re := regexp.MustCompile(dbdumpReg)
+	re := regexp.MustCompile(`(["])(.*?\.(rar|sql.gz))"`)
 	dbdumps := re.FindAllString(string(response), -1)
 
 	for i, dbdump := range dbdumps {
@@ -233,34 +272,6 @@ func ParseDbdumps(response []byte) []string {
 	}
 
 	return dbdumps
-}
-
-func getBody(baseURL string) ([]byte, error) {
-	client := http.Client{
-		Timeout: HTTPClientTimeout,
-		Transport: &http.Transport{
-			Proxy:           http.ProxyFromEnvironment,
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		}}
-	r, err := client.Get(baseURL)
-	if err != nil {
-		log.Printf("http.Get(%q) error: %v", baseURL, err)
-		return nil, err
-	}
-	if r.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("unable to reach to mirror %v: %v", baseURL, r.StatusCode)
-	}
-
-	b, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := r.Body.Close(); err != nil {
-		return nil, err
-	}
-
-	return b, nil
 }
 
 // parseHashes takes in a HTTP response and scans it for
@@ -327,61 +338,6 @@ func parseResponse(response []byte) (*Book, error) {
 	}
 
 	return &book, nil
-}
-
-func printDetails(book *Book) error {
-	var fsize string
-	size, err := strconv.Atoi(book.Filesize)
-	if err != nil {
-		fsize = "N/A"
-	} else {
-		fsize = humanize.Bytes(uint64(size))
-	}
-
-	// Print separation lines
-	fmt.Println(strings.Repeat("-", 80))
-
-	// Print ID + Title
-	fTitle := fmt.Sprintf("%5s %s", color.New(color.FgHiBlue).Sprintf(book.ID), book.Title)
-	fTitle = formatTitle(fTitle, TitleMaxLength)
-	if runtime.GOOS == "windows" {
-		_, err = fmt.Fprintf(color.Output, "%s\n    ++ ", fTitle)
-		if err != nil {
-			return err
-		}
-	} else {
-		fmt.Printf("%s\n    ++ ", fTitle)
-	}
-
-	// Slice author name if it exceeds AuthorMaxLength
-	var formatAuthor string
-	if len(book.Author) > AuthorMaxLength {
-		formatAuthor = book.Author[:AuthorMaxLength]
-	} else if book.Author == "" {
-		formatAuthor = "N/A"
-	} else {
-		formatAuthor = book.Author
-	}
-
-	err = prettify("author", formatAuthor, color.FgYellow, "-25")
-	if err != nil {
-		return err
-	}
-	err = prettify("year", book.Year, color.FgCyan, "4")
-	if err != nil {
-		return err
-	}
-	err = prettify("size", fsize, color.FgGreen, "6")
-	if err != nil {
-		return err
-	}
-	err = prettify("type", book.Extension, color.FgRed, "4")
-	if err != nil {
-		return err
-	}
-	fmt.Println()
-
-	return nil
 }
 
 // formatTitle shortens the title of a Book down to
