+++ libgen/download.go
@@ -16,11 +16,11 @@
 package libgen
 
 import (
-	"crypto/tls"
 	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
+	"log"
 	"math/rand"
 	"net/http"
 	"net/url"
@@ -57,17 +57,43 @@ func DownloadBook(book *Book, outputPath string) error {
 		filesize = r.ContentLength
 		bar := pb.Full.Start64(filesize)
 
-		out, err := makeFile(outputPath, filename)
-		if err != nil {
-			return err
+		var out *os.File
+		var mkErr error
+		// if output path was not provided
+		if outputPath == "" {
+			wd, err := os.Getwd()
+			if err != nil {
+				return err
+			}
+			if stat, err := os.Stat(fmt.Sprintf("%s/libgen", wd)); err == nil && stat.IsDir() {
+				out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
+			} else {
+				if err := os.Mkdir(fmt.Sprintf("%s/libgen", wd), 0755); err != nil {
+					return err
+				}
+				out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
+			}
+			if mkErr != nil {
+				return mkErr
+			}
+			// If output path was provided
+		} else {
+			if stat, err := os.Stat(outputPath); err == nil && stat.IsDir() {
+				out, err = os.Create(fmt.Sprintf("%s/%s", outputPath, filename))
+				if err != nil {
+					return err
+				}
+			} else {
+				return errors.New("invalid output path")
+			}
 		}
+
 		_, err = io.Copy(out, bar.NewProxyReader(r.Body))
 		if err != nil {
 			return err
 		}
 
 		bar.Finish()
-
 		if err := out.Close(); err != nil {
 			return err
 		}
@@ -96,10 +122,37 @@ func DownloadDbdump(filename string, outputPath string) error {
 		filesize := r.ContentLength
 		bar := pb.Full.Start64(filesize)
 
-		out, err := makeFile(outputPath, filename)
-		if err != nil {
-			return err
+		var out *os.File
+		var mkErr error
+		// if output path was not provided
+		if outputPath == "" {
+			wd, err := os.Getwd()
+			if err != nil {
+				return err
+			}
+			if stat, err := os.Stat(fmt.Sprintf("%s/libgen", wd)); err == nil && stat.IsDir() {
+				out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
+			} else {
+				if err := os.Mkdir(fmt.Sprintf("%s/libgen", wd), 0755); err != nil {
+					return err
+				}
+				out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
+			}
+			if mkErr != nil {
+				return mkErr
+			}
+			// If output path was provided
+		} else {
+			if stat, err := os.Stat(outputPath); err == nil && stat.IsDir() {
+				out, err = os.Create(fmt.Sprintf("%s/%s", outputPath, filename))
+				if err != nil {
+					return err
+				}
+			} else {
+				return errors.New("invalid output path")
+			}
 		}
+
 		_, err = io.Copy(out, bar.NewProxyReader(r.Body))
 		if err != nil {
 			return err
@@ -126,40 +179,31 @@ func DownloadDbdump(filename string, outputPath string) error {
 func GetDownloadURL(book *Book) error {
 	chosenMirror := DownloadMirrors[rand.Intn(3)]
 
-	var x int
-	tries := 3
-	for tries >= x {
-		switch chosenMirror.String() {
-		case "80.82.78.13":
-			if err := getBooksdlDownloadURL(book); err != nil {
-				if err = getBokDownloadURL(book); err != nil {
-					if err := getNineThreeURL(book); err != nil {
-						return err
-					}
+	switch chosenMirror.String() {
+	case "80.82.78.13":
+		if err := getBooksdlDownloadURL(book); err != nil {
+			if err = getBokDownloadURL(book); err != nil {
+				if err := getNineThreeURL(book); err != nil {
+					return err
 				}
 			}
-		case "https://b-ok.cc":
-			if err := getBokDownloadURL(book); err != nil {
-				if err = getNineThreeURL(book); err != nil {
-					if err = getBooksdlDownloadURL(book); err != nil {
-						return err
-					}
-				}
-			}
-		case "http://93.174.95.29":
-			if err := getNineThreeURL(book); err != nil {
+		}
+	case "https://b-ok.cc":
+		if err := getBokDownloadURL(book); err != nil {
+			if err = getNineThreeURL(book); err != nil {
 				if err = getBooksdlDownloadURL(book); err != nil {
-					if err = getBokDownloadURL(book); err != nil {
-						return err
-					}
+					return err
 				}
 			}
 		}
-		if book.DownloadURL != "" {
-			break
+	case "http://93.174.95.29":
+		if err := getNineThreeURL(book); err != nil {
+			if err = getBooksdlDownloadURL(book); err != nil {
+				if err = getBokDownloadURL(book); err != nil {
+					return err
+				}
+			}
 		}
-		// Increment tries
-		x++
 	}
 
 	if book.DownloadURL == "" {
@@ -179,12 +223,26 @@ func getBooksdlDownloadURL(book *Book) error {
 	baseURL.RawQuery = q.Encode()
 	book.PageURL = baseURL.String()
 
-	b, err := getBody(baseURL.String())
+	client := http.Client{Timeout: HTTPClientTimeout, Transport: &http.Transport{Proxy: http.ProxyFromEnvironment}}
+	r, err := client.Get(baseURL.String())
+	if err != nil {
+		log.Printf("http.Get(%q) error: %v", baseURL, err)
+		return err
+	}
+	if r.StatusCode != http.StatusOK {
+		return fmt.Errorf("unable to reach to mirror %v: %v", baseURL.Host, r.StatusCode)
+	}
+
+	b, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		return err
 	}
 
 	book.DownloadURL = string(findMatch(booksdlReg, b))
+
+	if err := r.Body.Close(); err != nil {
+		return err
+	}
 
 	return nil
 }
@@ -198,7 +256,17 @@ func getBokDownloadURL(book *Book) error {
 	queryURL := baseURL.String() + book.Md5
 	book.PageURL = queryURL
 
-	b, err := getBody(queryURL)
+	client := http.Client{Timeout: HTTPClientTimeout, Transport: &http.Transport{Proxy: http.ProxyFromEnvironment}}
+	resp, err := client.Get(queryURL)
+	if err != nil {
+		return err
+	}
+
+	if resp.StatusCode != http.StatusOK {
+		return fmt.Errorf("unable to reach to mirror %v: %v", baseURL.Host, resp.StatusCode)
+	}
+
+	b, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		return err
 	}
@@ -211,6 +279,10 @@ func getBokDownloadURL(book *Book) error {
 	book.DownloadURL = "https://b-ok.cc" + string(downloadURL)
 
 	if err := checkBokDownloadLimit(book); err != nil {
+		return err
+	}
+
+	if err := resp.Body.Close(); err != nil {
 		return err
 	}
 
@@ -227,12 +299,7 @@ func checkBokDownloadLimit(book *Book) error {
 		return err
 	}
 	req.Header.Add("Referer", book.PageURL)
-	client := http.Client{
-		Timeout: HTTPClientTimeout,
-		Transport: &http.Transport{
-			Proxy:           http.ProxyFromEnvironment,
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		}}
+	client := http.Client{Timeout: HTTPClientTimeout, Transport: &http.Transport{Proxy: http.ProxyFromEnvironment}}
 	resp, err := client.Do(req)
 	if err != nil {
 		return err
@@ -265,7 +332,17 @@ func getNineThreeURL(book *Book) error {
 	queryURL := baseURL.String() + book.Md5
 	book.PageURL = queryURL
 
-	b, err := getBody(queryURL)
+	client := http.Client{Timeout: HTTPClientTimeout, Transport: &http.Transport{Proxy: http.ProxyFromEnvironment}}
+	resp, err := client.Get(queryURL)
+	if err != nil {
+		return err
+	}
+
+	if resp.StatusCode != http.StatusOK {
+		return fmt.Errorf("unable to reach to mirror %v: %v", baseURL.Host, resp.StatusCode)
+	}
+
+	b, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		return err
 	}
@@ -277,48 +354,11 @@ func getNineThreeURL(book *Book) error {
 
 	book.DownloadURL = "http://93.174.95.29" + string(downloadURL)
 
+	if err := resp.Body.Close(); err != nil {
+		return err
+	}
+
 	return nil
-}
-
-func makeFile(outputPath, filename string) (*os.File, error) {
-	var out *os.File
-	var mkErr error
-
-	// Handle long titles
-	if len(filename) >= 256 {
-		filename = filename[:256]
-	}
-
-	// if output path was not provided
-	if outputPath == "" {
-		wd, err := os.Getwd()
-		if err != nil {
-			return nil, err
-		}
-		if stat, err := os.Stat(fmt.Sprintf("%s/libgen", wd)); err == nil && stat.IsDir() {
-			out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
-		} else {
-			if err := os.Mkdir(fmt.Sprintf("%s/libgen", wd), 0755); err != nil {
-				return nil, err
-			}
-			out, mkErr = os.Create(fmt.Sprintf("%s/libgen/%s", wd, filename))
-		}
-		if mkErr != nil {
-			return nil, mkErr
-		}
-		// If output path was provided
-	} else {
-		if stat, err := os.Stat(outputPath); err == nil && stat.IsDir() {
-			out, err = os.Create(fmt.Sprintf("%s/%s", outputPath, filename))
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			return nil, errors.New("invalid output path")
-		}
-	}
-
-	return out, nil
 }
 
 // findMatch is a helper function that searches an []byte
