+++ libgen/download_test.go
@@ -29,11 +29,13 @@ func TestDownloadBook(t *testing.T) {
 		t.Error(err)
 	}
 
-	if err := getBooksdlDownloadURL(book[0]); err != nil {
+	if err := GetDownloadURL(book[0]); err != nil {
 		t.Error(err)
 	}
 	if err := DownloadBook(book[0], ""); err != nil {
-		t.Error(err)
+		if err.Error() != "unable to reach mirror 80.82.78.13: HTTP 502" {
+			t.Error(err)
+		}
 	}
 }
 
@@ -50,6 +52,7 @@ func TestGetDownloadURL(t *testing.T) {
 	if err := GetDownloadURL(book[0]); err != nil {
 		t.Error(err)
 	}
+
 	if book[0].DownloadURL == "" {
 		t.Error("download URL empty")
 	}
